requirements.txt
fastapi
uvicorn[standard]
transformers
torch
sentence-transformers
faiss-cpu
numpy
pydantic
httpx
redis
aiohttp
beautifulsoup4
lxml
tenacity
chardet
